Improved performance by emulating the sprite cache

+++ src/disp.h
@@ -14,7 +14,7 @@ private:
 	void clear();
 	uint32_t colors[4];
 	double rem = 0;
-	uint32_t frame_time = 0;
+	uint64_t frame_time = 0;
 	bool poll(State* ctx);
 public:
 	bool init(State* ctx, int scale_factor);
